requirements.txt
Pillow
watchdog
pyinstaller
pyserial
